refactor: перенёс login/send_batch/flush_and_send в core.py, добавил тесты на фильтрацию/буфер/очередь (часть #13)
